move PaginatedQueryHandlerBase to VoteMonitor.Api.Core.Extensions and rename to QueryExtension
